rest-example/requirements.txt
flask==2.3.2
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
mysqlclient==2.2.0
psycopg2==2.9.6